[Update] Redefine gapi endpoints
