+++ .gitignore
@@ -3,3 +3,4 @@
 __pycache__/
 token.json
 .env
+mock_data/
