+++ app/__init__.py
@@ -1,9 +1,9 @@
-from fastapi import FastAPI, Request, Depends
+from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
 from app.api.v1.project import project_router
 from app.api.v1.ctscan import ctscan_router
-from app.api.v1.gsheets import gsheets_router
+from app.api.v1.gapi import gapi_router
 from app.core.db.pgsql.session import engine
 from app.core.db.pgsql import Base
 from app.models import *
@@ -26,8 +26,8 @@ def init_cors(app: FastAPI) -> None:
 
 def init_routers(app: FastAPI) -> None:
     app.include_router(project_router)
+    app.include_router(gapi_router)
     app.include_router(ctscan_router)
-    app.include_router(gsheets_router)
 
 
 def create_app() -> FastAPI:
