+++ app/api/v1/ctscan/dal.py
@@ -1,0 +1,11 @@
+from typing import List, Optional, Union
+from sqlalchemy.future import select
+from sqlalchemy import update, delete
+from sqlalchemy.orm import Session
+
+from app.core.db.pgsql import async_session
+
+
+class CTscanDAL:
+    def __init(self) -> None:
+        pass
