+++ app/api/v1/gapi/__init__.py
@@ -1,0 +1,1 @@
+from .endpoints import gapi_router
