+++ app/api/v1/gapi/dal.py
@@ -7,20 +7,20 @@ class GSheetsDAL:
     def __init__(self) -> None:
         self.qctworksheet = QCTWorksheet().sheet
 
-    def get_projects_list(self):
+    def get_project_list(self):
         worksheet_list = self.qctworksheet.worksheets()
         return worksheet_list
 
     def get_project_data(self, project: str):
         project_worksheet = self.qctworksheet.worksheet(project)
         project_data = project_worksheet.get_all_values()
+
         if isinstance(project_data, List) and len(project_data) > 0:
             headers = project_data[0]
             rows = project_data[1:]
 
             project_headers = [
-                {"Header": header.upper(), "accessor": header}
-                for header in project_data[0]
+                {"Header": header.upper(), "accessor": header} for header in headers
             ]
             project_rows = [
                 {header: val for header, val in zip(headers, row)} for row in rows
