--- app/api/v1/gsheets/__init__.py
@@ -1,1 +1,0 @@
-from .endpoints import gsheets_router
