+++ requirements.txt
@@ -1,0 +1,54 @@
+anyio==3.5.0
+asgiref==3.5.0
+asyncpg==0.25.0
+black==22.3.0
+cachetools==5.0.0
+certifi==2021.10.8
+charset-normalizer==2.0.12
+click==8.0.3
+distlib==0.3.4
+fastapi==0.73.0
+filelock==3.6.0
+google-api-core==2.5.0
+google-api-python-client==2.37.0
+google-auth==2.6.0
+google-auth-httplib2==0.1.0
+google-auth-oauthlib==0.4.6
+googleapis-common-protos==1.54.0
+greenlet==1.1.2
+gspread==5.3.2
+h11==0.13.0
+httplib2==0.20.4
+httptools==0.3.0
+idna==3.3
+loguru==0.6.0
+mypy-extensions==0.4.3
+oauthlib==3.2.0
+pathspec==0.9.0
+pipenv==2022.3.24
+pkg_resources==0.0.0
+platformdirs==2.5.1
+protobuf==3.19.4
+pyasn1==0.4.8
+pyasn1-modules==0.2.8
+pydantic==1.9.0
+pyparsing==3.0.7
+python-dotenv==0.19.2
+PyYAML==6.0
+requests==2.27.1
+requests-oauthlib==1.3.1
+rsa==4.8
+six==1.16.0
+sniffio==1.2.0
+SQLAlchemy==1.4.35
+starlette==0.17.1
+tomli==2.0.1
+typing_extensions==4.0.1
+uritemplate==4.1.1
+urllib3==1.26.8
+uvicorn==0.17.1
+uvloop==0.16.0
+virtualenv==20.13.4
+virtualenv-clone==0.5.7
+watchgod==0.7
+websockets==10.1
